requirements.txt
anyio==4.4.0
asgiref==3.7.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
certifi==2024.2.2
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cycler==0.12.1
dj-database-url==2.1.0
Django==4.2.1
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
fonttools==4.53.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
itypes==1.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
numpy==2.0.1
packaging==23.2
pillow==10.3.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
whitenoise==6.7.0
wsproto==1.2.0
zope.interface==7.0.3
